+++ integrations/amazon_bedrock/CHANGELOG.md
@@ -1,5 +1,24 @@
 # Changelog
 
+## [integrations/amazon_bedrock-v3.0.0] - 2025-01-23
+
+### 🚀 Features
+
+- *(AWS Bedrock)* Add Cohere Reranker (#1291)
+- AmazonBedrockChatGenerator - add tools support (#1304)
+
+### 🚜 Refactor
+
+- [**breaking**] AmazonBedrockGenerator - remove truncation  (#1314)
+
+
+## [integrations/amazon_bedrock-v2.1.3] - 2025-01-21
+
+### 🧹 Chores
+
+- Bedrock - pin `transformers!=4.48.*` (#1306)
+
+
 ## [integrations/amazon_bedrock-v2.1.2] - 2025-01-20
 
 ### 🌀 Miscellaneous