+++ integrations/google_ai/CHANGELOG.md
@@ -1,5 +1,26 @@
 # Changelog
 
+## [integrations/google_ai-v5.1.0] - 2025-01-30
+
+### 🚀 Features
+
+- Google AI Chat Generator - add tool config + improvements (#1333)
+
+
+## [integrations/google_ai-v5.0.1] - 2025-01-23
+
+### 🚜 Refactor
+
+- GoogleAIGeminiGenerator - make some attributes public (#1317)
+
+
+## [integrations/google_ai-v5.0.0] - 2025-01-23
+
+### 🚀 Features
+
+- [**breaking**] Google AI - support for Tool + general refactoring (#1316)
+
+
 ## [integrations/google_ai-v4.1.0] - 2025-01-16
 
 ### 🧹 Chores